# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Ikigai Compass API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /obstacles:
    get:
      operationId: listObstacles
      tags: [obstacles]
      summary: List obstacles
      responses:
        "200":
          description: List of obstacles
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Obstacle"
    post:
      operationId: createObstacle
      tags: [obstacles]
      summary: Log an obstacle
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateObstacleBody"
      responses:
        "201":
          description: Created obstacle
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Obstacle"
  /obstacles/{id}:
    put:
      operationId: updateObstacle
      tags: [obstacles]
      summary: Update obstacle status
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateObstacleBody"
      responses:
        "200":
          description: Updated obstacle
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Obstacle"
...
  /obstacles/{id}/help:
    post:
      tags: [Obstacles]
      summary: Get AI-generated help for overcoming an obstacle
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ObstacleHelpBody"
      responses:
        "200":
          description: AI help for the obstacle
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ObstacleHelpResponse"
        "404":
          description: Obstacle not found
...
  /ai/extract-themes:
    post:
      tags: [AI]
      summary: Extract themes from a block of text using AI
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ExtractThemesBody"
      responses:
        "200":
          description: Extracted themes
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ExtractThemesResponse"
  /ai/generate-insights:
    post:
      tags: [AI]
      summary: Generate insight cards from a discovery context using AI
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GenerateInsightsBody"
      responses:
        "200":
          description: Generated insight cards
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GenerateInsightsResponse"
...
  /insights:
    get:
      operationId: listInsights
      tags: [Insights]
      summary: List all insights for the current user
      security:
        - bearerAuth: []
      responses:
        "200":
          description: List of insights
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Insight"
    post:
      operationId: createInsight
      tags: [Insights]
      summary: Create a new insight
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateInsightBody"
      responses:
        "201":
          description: Created insight
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Insight"
  /insights/{id}:
    get:
      operationId: getInsight
      tags: [Insights]
      summary: Get a specific insight
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Insight details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Insight"
        "404":
          description: Not found
    put:
      operationId: updateInsight
      tags: [Insights]
      summary: Update an insight
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateInsightBody"
      responses:
        "200":
          description: Updated insight
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Insight"
    delete:
      operationId: deleteInsight
      tags: [Insights]
      summary: Delete an insight
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /themes:
    get:
      operationId: listThemes
      tags: [Insights]
      summary: List all themes for the current user
      security:
        - bearerAuth: []
      responses:
        "200":
          description: List of themes
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Theme"
    post:
      operationId: createTheme
      tags: [Insights]
      summary: Create a new theme
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateThemeBody"
      responses:
        "201":
          description: Created theme
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Theme"
  /themes/{id}:
    put:
      operationId: updateTheme
      tags: [Insights]
      summary: Update a theme
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateThemeBody"
      responses:
        "200":
          description: Updated theme
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Theme"
    delete:
      operationId: deleteTheme
      tags: [Insights]
      summary: Delete a theme
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
...
components:
  schemas:
...
    Obstacle:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: integer
        planId:
          type: integer
        title:
          type: string
        description:
          type: string
        category:
          type: string
          enum: [time, energy, skills, resources, motivation, external, other]
        status:
          type: string
          enum: [active, resolved, accepted]
        aiSuggestions:
          type: string
        resolvedAt:
          type: string
          format: date-time
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - userId
        - title
        - category
        - status
        - createdAt
        - updatedAt
    CreateObstacleBody:
      type: object
      properties:
        planId:
          type: integer
        title:
          type: string
        description:
          type: string
        category:
          type: string
          enum: [time, energy, skills, resources, motivation, external, other]
      required:
        - title
        - category
    UpdateObstacleBody:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        category:
          type: string
          enum: [time, energy, skills, resources, motivation, external, other]
        status:
          type: string
          enum: [active, resolved, accepted]
...
    ObstacleHelpBody:
      type: object
      properties:
        context:
          type: string
    ObstacleHelpResponse:
      type: object
      properties:
        reframe:
          type: string
        rootCauses:
          type: array
          items:
            type: string
        strategies:
          type: array
          items:
            type: string
        reflectionQuestions:
          type: array
          items:
            type: string
        isAiGenerated:
          type: boolean
      required:
        - reframe
        - rootCauses
        - strategies
        - reflectionQuestions
        - isAiGenerated
...
    ExtractThemesBody:
      type: object
      properties:
        text:
          type: string
        sessionId:
          type: integer
      required:
        - text
    ExtractThemesResponse:
      type: object
      properties:
        themes:
          type: array
          items:
            type: object
            properties:
              name:
                type: string
              description:
                type: string
              keywords:
                type: array
                items:
                  type: string
        isAiGenerated:
          type: boolean
      required:
        - themes
        - isAiGenerated
    GenerateInsightsBody:
      type: object
      properties:
        context:
          type: string
        sessionId:
          type: integer
        hypothesisId:
          type: integer
      required:
        - context
    GenerateInsightsResponse:
      type: object
      properties:
        insights:
          type: array
          items:
            type: object
            properties:
              title:
                type: string
              content:
                type: string
              category:
                type: string
        isAiGenerated:
          type: boolean
      required:
        - insights
        - isAiGenerated
...
    Insight:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: integer
        title:
          type: string
        content:
          type: string
        category:
          type: string
        sourceType:
          type: string
        isAiGenerated:
          type: boolean
        createdAt:
          type: string
        updatedAt:
          type: string
      required:
        - id
        - userId
        - title
        - content
        - createdAt
        - updatedAt
    CreateInsightBody:
      type: object
      properties:
        title:
          type: string
        content:
          type: string
        category:
          type: string
        sourceType:
          type: string
      required:
        - title
        - content
    UpdateInsightBody:
      type: object
      properties:
        title:
          type: string
        content:
          type: string
        category:
          type: string
    Theme:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: integer
        name:
          type: string
        description:
          type: string
        frequency:
          type: integer
        isAiExtracted:
          type: boolean
        createdAt:
          type: string
        updatedAt:
          type: string
      required:
        - id
        - userId
        - name
        - createdAt
        - updatedAt
    CreateThemeBody:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        frequency:
          type: integer
      required:
        - name
    UpdateThemeBody:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        frequency:
          type: integer
...